requirements.txt
flask
google-cloud-logging
gunicorn
msgpack
numpy
onnxruntime
pillow
pdfplumber
psutil
pypdf2
pypdfium2
reportlab
tqdm
